Map Maker 0.2 Screen scrolling and height changing
(removed the leftmost digit from version number to avoid confusion with
main game version numbers) You can now move about the screen freely.
Main update is that clicking on a tile results in an event. Right now
it's just height changing but it paves the way for the rest of the
functions.